first

+++ index.js
@@ -2,6 +2,10 @@ const express = require("express");
 const bodyParser = require("body-parser");
 const cors = require("cors");
 const mongoose = require("mongoose");
+const bcrypt = require("bcrypt");
+const session = require("express-session");
+const passport = require("passport");
+const passportLocal = require("passport-local");
 
 mongoose.connect(
   "mongodb+srv://ajdb:Talofa123@cluster0-a37uh.mongodb.net/herodatabase?retryWrites=true&w=majority",
@@ -11,38 +15,156 @@ mongoose.connect(
   }
 );
 
+var userSchema = new mongoose.Schema({
+  name: {
+    type: String,
+    required: true
+  },
+  email: {
+    type: String,
+    required: true
+  },
+  encryptedPassword: {
+    type: String,
+    required: true
+  }
+});
+
+userSchema.methods.setEcryptedPassword = function(
+  plainPassword,
+  callbackFunction
+) {
+  //this is the user instance
+  bcrypt.hash(plainPassword, 12).then(hash => {
+    this.encryptedPassword = hash;
+    callbackFunction();
+  });
+};
+
+userSchema.methods.verifyPassword = function(plainPassword, callbackFunction) {
+  bcrypt.compare(plainPassword, this.encryptedPassword).then(result => {
+    callbackFunction(result);
+  });
+};
+
+var User = mongoose.model("User", userSchema);
+
 var Dog = mongoose.model("Dog", {
   name: {
-    type:String, 
+    type: String,
     requred: [true, "Name required"]
   },
   breed: {
-    type:String, 
+    type: String,
     requred: [true, "Breed required"]
   },
   age: {
-    type:String, 
+    type: String,
     requred: [true, "Age required"]
   },
   url: {
-    type:String, 
+    type: String,
     requred: [true, "URL required"]
   }
 });
 
 const app = express();
 const port = process.env.PORT || 3000;
+
 app.use(bodyParser.urlencoded({ extended: false }));
-app.use(cors());
+app.use(cors({ credentials: true, origin: "null" }));
+app.use(
+  session({
+    secret: "jytfgvjytf98795445676",
+    resave: false,
+    saveUninitialized: true
+  })
+);
+app.use(passport.initialize());
+app.use(passport.session());
+
+passport.use(
+  new passportLocal.Strategy(
+    {
+      usernameField: "email",
+      passwordField: "plainPassword"
+    },
+    function(email, plainPassword, done) {
+      User.findOne({ email: email })
+        .then(function(user) {
+          if (!user) {
+            return done(null, false);
+          } else {
+            user.verifyPassword(plainPassword, function(result) {
+              if (result) {
+                return done(null, user);
+              } else {
+                return done(null, false);
+              }
+            });
+          }
+        })
+        .catch(function(err) {
+          done(err);
+        });
+    }
+  )
+);
+
+passport.serializeUser(function(user, done) {
+  done(null, user._id);
+});
+
+passport.deserializeUser(function(id, done) {
+  User.findOne({ _id: id }).then(function(user) {
+    done(null, user);
+  });
+});
+
+app.post("/session", passport.authenticate("local"), function(req, res) {
+  console.log("user", req.user);
+  res.sendStatus(201);
+});
+
+app.get("/session", function(req, res) {
+  if (req.user) {
+    res.json(req.user);
+  } else {
+    res.sendStatus(401);
+    console.log("no usr");
+  }
+});
+
+app.post("/users", function(req, res) {
+  // Store hash in your password DB.
+  let user = new User({
+    name: req.body.name,
+    email: req.body.email
+  });
+  user.setEcryptedPassword(req.body.plainPassword, function() {
+    user
+      .save()
+      .then(function() {
+        res.sendStatus(201);
+      })
+      .catch(function(err) {
+        console.log(err);
+      });
+  });
+});
 
 app.get("/dogs", function(req, res) {
-  console.log("GET called");
+  //console.log("user", req.user);
   Dog.find({}).then(function(dogs) {
     res.json(dogs);
   });
 });
 
 app.post("/dogs", function(req, res) {
+  if (!req.user) {
+    res.sendStatus(401);
+    return;
+  }
   console.log("Posted the body", req.body);
   req.body.name;
 
@@ -59,6 +181,10 @@ app.post("/dogs", function(req, res) {
 });
 
 app.delete("/dogs/:dogId", function(req, res) {
+  if (!req.user) {
+    res.sendStatus(401);
+    return;
+  }
   console.log("DELETE Called");
   let dogId = req.params.dogId;
   Dog.findOneAndDelete({ _id: dogId })
@@ -75,6 +201,10 @@ app.delete("/dogs/:dogId", function(req, res) {
 });
 
 app.put("/dogs/:dogId", function(req, res) {
+  if (!req.user) {
+    res.sendStatus(401);
+    return;
+  }
   console.log("Put on server called");
   console.log(req.body);
   let dogId = req.params.dogId;
@@ -83,7 +213,7 @@ app.put("/dogs/:dogId", function(req, res) {
       dog.name = req.body.name;
       dog.age = req.body.age;
       dog.breed = req.body.breed;
-      dog.url = req.body.url
+      dog.url = req.body.url;
       dog.save().then(function() {
         res.sendStatus(202);
       });
